Update lsp implementation to 0.11
Neovim v0.11 introduces `vim.lsp.enable` and `vim.lsp.config` which has
the capability of fully replacing `nvim-lspconfig`. This update migrates
to using the native LSP implementation, with dynamic-enabling LSPs based
on the existance of a configuration in the `lsp/` directory.

Other changes include refactoring a few other plugins and their
configurations.